refactor: InputString을 MVC패턴을 적용하기 위해 InputView로 수정
