+++ src/main/java/racingcar/Application.java
@@ -1,14 +1,13 @@
 package racingcar;
 
-import static racingcar.Input.InputString.inputString;
-
 import camp.nextstep.edu.missionutils.Console;
+import racingcar.view.InputView;
 
 public class Application {
 
     public static void main(String[] args) {
-        String input = Console.readLine();
-        System.out.println(inputString(input));
+
+        InputView InputView = new InputView();
     }
